requirements.txt
openai
fastapi
uvicorn
jinja2
python-multipart
pydantic-settings
streamlit